[RFC] Introduce @Final DynamicItem for immutable bindings
In some cases plugins would like to expose their API to other plugins,
for example to give access to internal state or to programatically
modify its configuraiton.

This can be potentially done currently with the usage of DynamicItem,
but that doesn't guarantee that the implemenation will be overwritten
by other plugins.

Introduce the DynamicItem.Final annotation that, once assigned to a type,
requires the type to be bound only once.

Any attempt to re-bind a DynamicItem.Final twice would result in
a Guice ProvisionException.

Bug: Issue 338786480
Release-Notes: Introduce @Final DynamicItem for allowing immutable dynamic bindings of interfaces to implementations in other plugins.
Change-Id: Ie7edf73e27fc6de898dbf248b9c4cda6b97e8153