Introduce @DynamicItem.Final for immutable bindings
In some cases plugins would like to expose their API to other plugins,
for example to give access to internal state or to programatically
modify its configuration.

This can be potentially done currently with the usage of DynamicItem,
but that doesn't guarantee that the implementation will not be
overwritten by other plugins.
Also, one plugin may have specific expectations of where the
implementation of those interfaces must be found and
therefore exposing that constraint to Gerrit plugin load so that
only the allowed plugin is accepted for binding that item.

Introduce the @DynamicItem.Final annotation that, once assigned to a
type, requires the type to be bound at most once.

Any attempt to re-bind a DynamicItem.Final twice or from a plugin
which is not the expected one, would result in a Guice
ProvisionException.

Example:

@DynamicItem.Final(implementedByPlugin="foo-plugin")
public interface FooInterface {
    String fooApi(String param);
}

Bug: Issue 338786480
Release-Notes: Introduce @DynamicItem.Final for allowing immutable dynamic bindings of interfaces to implementations in other plugins.
Change-Id: Ie7edf73e27fc6de898dbf248b9c4cda6b97e8153