Fix reloading of plugins registering dynamic items
When change Ia7ed2a06ec3 refactored the `Plugin` hierarchy by
introducing different plugin specifications, omitted to inherit the
`reloadableHandles` from the parent class and instead, introduced a new
`reloadableHandles` variable in the `JarPlugin` implementation.

This caused the `Plugin.getReloadableHandles()` to always return an
empty list, effectively preventing the reloading of the plugin from
replacing existing reloadable bindings [1] and creating new ones
instead.

Whilst this still allows the `reload` to succeed, it misses the removal
of the registration handle associated to the old plugin instance, which
will then not be de-registered if the new plugin is then disabled later
on.

Inherit `reloadableHandles` from the parent `Plugin` class so that
addition and retrieval of reloadable registration handles targets the
same instance variable, allowing the reload to move existing dynamic
registration handles to the new plugin instance, rather than creating a
new one.

[1] https://gerrit.googlesource.com/gerrit/+/refs/heads/stable-3.8/java/com/google/gerrit/server/plugins/PluginGuiceEnvironment.java#305

Bug: Issue 346618790
Release-Notes: Fix reload of plugins causing incorrect deregistration of dynamic items
Change-Id: I13896cf1fcabebc26d465aec7c8debc9dddb11be