Expose whether project state predicate is enabled
When implementing filtering of read-only repositories for the web view
`/admin/repos`, we found out Google disabled querying the state of
project in the REST API `/projects/`.

This introduce `gerrit.projectStatePredicateEnabled` in the config with
a default of true and exposes it in the server info:

  GET /config/server/info
  {
    "gerrit": {
      "all_projects": "All-Projects",
      "all_users": "All-Users",
      "doc_search": false,
      "project_state_predicate_enabled": true
    }
  ...

This would let the frontend find out whether it can or can not filter
project by state.

Release-Notes: skip
Change-Id: Id2eff9a9851004af89224cf5547383c2c7098e98