Make explicit the Gerrit support for Windows Server
Gerrit documentation has several references on how to setup
the service on Windows Server, making the reader thinking it as
a valid option for a production setup.

Gerrit isn't actively tested on the CI/CD on Windows Server
and its releases not validated end-to-end. There is also
a very limited adoption from the community therefore the
feedback on potential issues is very limited.

Add an explicit note on the current situation with Windows
Server so that Gerrit admins are warned on the risks associated.

Change-Id: I408c4a8ec3f2a3aa4552b5d8cbfcd981c17e5b89