Fix a special case for preview/apply fix.
When a file doesn't have end of line mark at the end, some fixes have the end at the line after the last line instead of last character in the last line and this caused an exception.

It seems this is a common case for some checkers/linters/etc..., so we should accept such input as well.

Google-Bug: b/362428689
Release-Notes: skip
Change-Id: Idfb99fed9148584c9b7eb83e6ac9216689bbb3da